transfer messenger server routes to messageRoutes used Router
