+++ server.js
@@ -1,141 +1,20 @@
 "use strict";
 const express = require("express");
-const bodyParser = require("body-parser");
-const app = express().use(bodyParser.json());
-const request = require("request");
-require("dotenv").config();
+const cors = require('cors');
 
-function handleMessage(senderPsid, receivedMessage) {
-  let response;
-  // Checks if the message contains text
-  if (receivedMessage.text) {
-    // Create the payload for a basic text message, which
-    // will be added to the body of our request to the Send API
-    response = {
-      // 'text': `You sent the message: "${receivedMessage.text}". Now send me an attachment!`
-      text: `Contrary to popular belief, Lorem Ipsum is not simply random text. It has roots in a piece of classical Latin literature from 45 BC, making it over 2000 years old. Richard McClintock, a Latin professor at Hampden-Sydney College in Virginia, looked up one of the more obscure Latin words, consectetur, from a Lorem Ipsum passage, and going through the cites of the word in classical literature, discovered the undoubtable source. Lorem Ipsum comes from sections 1.10.32 and 1.10.33 of "de Finibus Bonorum et Malorum" (The Extremes of Good and Evil) by Cicero, written in 45 BC. This book is a treatise on the theory of ethics, very popular during the Renaissance. The first line of Lorem Ipsum, "Lorem ipsum dolor sit amet..", comes from a line in section 1.10.32.
-        The standard chunk of Lorem Ipsum used since the 1500s is reproduced below for those interested. Sections 1.10.32 and 1.10.33 from "de Finibus Bonorum et Malorum" by Cicero are also reproduced in their exact original form, accompanied by English versions from the 1914 translation by H. Rackham.
-        Where can I get some?
-       There are many variations of passages of Lorem Ipsum available, but the majority have suffered alteration in some form, by injected humour, or randomised words which don't look even slightly believable. If you are going to use a passage of Lorem Ipsum, you need to be sure there isn't anything embarrassing hidden in the middle of text. All the Lorem Ipsum generators on the Internet tend to repeat predefined chunks as necessary, making this the first true generator on the Internet. It uses a dictionary of over 200 Latin words, combined with a handful of model sentence structures, to generate Lorem Ipsum which looks reasonable. The generated Lorem Ipsum is therefore always free from repetition, injected humour, or non-characteristic words etc.`
-    };
-  } else if (receivedMessage.attachments) {
-    // Get the URL of the message attachment
-    let attachmentUrl = receivedMessage.attachments[0].payload.url;
-    response = {
-      attachment: {
-        type: "template",
-        payload: {
-          template_type: "generic",
-          elements: [
-            {
-              title: "Is this the right picture?",
-              subtitle: "Tap a button to answer.",
-              image_url: attachmentUrl,
-              buttons: [
-                {
-                  type: "postback",
-                  title: "Yes!",
-                  payload: "yes"
-                },
-                {
-                  type: "postback",
-                  title: "No!",
-                  payload: "no"
-                }
-              ]
-            }
-          ]
-        }
-      }
-    };
-  }
-  // Send the response message
-  callSendAPI(senderPsid, response);
-}
+// route declarations
+const messageRouter = require('./routes/messageRoutes');
 
-function handlePostback(senderPsid, receivedPostback) {
-  let response;
-  let payload = receivedPostback.payload;
-  if (payload === "yes") {
-    response = { text: "Thanks!" };
-  } else if (payload === "no") {
-    response = { text: "Oops, try sending another image." };
-  }
-  callSendAPI(senderPsid, response);
-}
+const server = express();
+server.use(cors());
+server.use(express.json());
 
-function callSendAPI(senderPsid, response) {
-  let requestBody = {
-    recipient: {
-      id: senderPsid
-    },
-    message: response
-  };
+// routes
+server.use('/api/messenger', messageRouter);
 
-  request(
-    {
-      uri: "https://graph.facebook.com/v2.6/me/messages",
-      qs: { access_token: process.env.PAGE_ACCESS_TOKEN },
-      method: "POST",
-      json: requestBody
-    },
-    (err, res, body) => {
-      if (!err) {
-        console.log("message sent");
-      } else {
-        console.error("unable to send message" + err);
-      }
-    }
-  );
-}
 
-app.post("/webhook", (req, res) => {
-  let body = req.body;
-  if (body.object === "page") {
-    body.entry.forEach(function(entry) {
-      let webhookEvent = entry.messaging[0];
-      console.log(webhookEvent);
-      // sender PSID
-      let senderPsid = webhookEvent.sender.id;
-      // reciever PSID
-      let receiverPsid = webhookEvent.recipient.id;
-
-      if (webhookEvent.message) {
-        handleMessage(senderPsid, webhookEvent.message);
-      } else if (webhookEvent.postback) {
-        handlePostback(senderPsid, webhookEvent.postback);
-      }
-    });
-    res.status(200).send("Event received");
-  } else {
-    res.sendStatus(404);
-  }
+server.get('/', (req, res) => {
+  res.status(200).json('Welcome to ChatReader server!');
 });
 
-app.get("/webhook", (req, res) => {
-  let VERIFY_TOKEN = process.env.VERIFY_TOKEN;
-  let mode = req.query["hub.mode"];
-  let token = req.query["hub.verify_token"];
-  let challenge = req.query["hub.challenge"];
-
-  if (mode && token) {
-    if (mode === "subscribe" && token === VERIFY_TOKEN) {
-      console.log("Webhook verified");
-      res.status(200).send(challenge);
-    } else {
-      res.sendStatus(403);
-    }
-  } else {
-    res.sendStatus(404);
-  }
-});
-
-app.get("/", (req, res) => {
-  res.send("API RUNNING!!!");
-  res.status(200);
-});
-
-const PORT = process.env.PORT || 5500;
-app.listen(PORT, () => {
-  console.log(`Server running on port ${PORT}`);
-});
+module.exports = server;
